pyproject.toml
[tool.poetry]
name = "article-summarizer"
version = "0.1.0"
description = ""
authors = ["kahnwong <karn@karnwong.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
readability-lxml = "^0.8.1"
summary = "^0.2.0"
lxml-html-clean = "^0.1.1"
beautifulsoup4 = "^4.12.3"
streamlit = "^1.33.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[tool.poetry.group.tests.dependencies]
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
